# Optimise the `BlackjackFunctional` `usable_ace` and `sum_hand` functions