--- a/gymnasium/envs/tabular/blackjack.py
+++ b/gymnasium/envs/tabular/blackjack.py
@@ -65,9 +65,10 @@ def draw_card(key, hand, index):
     return key, hand, index + 1
 
 
+@jax.jit
 def usable_ace(hand):
     """Checks to se if a hand has a usable ace."""
-    return jnp.logical_and((jnp.count_nonzero(hand == 1) > 0), (sum(hand) + 10 <= 21))
+    return jnp.logical_and(jnp.any(hand == 1), jnp.sum(hand) + 10 <= 21)
 
 
 def take(env_state):
@@ -134,9 +135,10 @@ def notake(env_state):
     )
 
 
+@jax.jit
 def sum_hand(hand):
     """Returns the total points in a hand."""
-    return sum(hand) + (10 * usable_ace(hand))
+    return jnp.sum(hand) + (10 * usable_ace(hand))
 
 
 def is_bust(hand):