--- a/gymnasium/envs/tabular/blackjack.py
+++ b/gymnasium/envs/tabular/blackjack.py
@@ -65,7 +65,6 @@ def draw_card(key, hand, index):
     return key, hand, index + 1
 
 
-@jax.jit
 def usable_ace(hand):
     """Checks to se if a hand has a usable ace."""
     return jnp.logical_and(jnp.any(hand == 1), jnp.sum(hand) + 10 <= 21)
@@ -135,7 +134,6 @@ def notake(env_state):
     )
 
 
-@jax.jit
 def sum_hand(hand):
     """Returns the total points in a hand."""
     return jnp.sum(hand) + (10 * usable_ace(hand))