@aseembits93 2025-08-11
Hi @pseudo-rnd-thoughts ! I undid the `jit` and ran the benchmark on 20 random episodes to get a ~20% speedup, I got the same speedup for up to 100 episodes. Here's the benchmark code - https://gist.github.com/aseembits93/2c0b232e93498a14e7686b0c8411ed10

Hope to hear back soon! Best,